+++ onadata/apps/logger/tests/models/test_instance.py
@@ -1306,97 +1306,3 @@ class TestInstance(TestBase):
         metadata.refresh_from_db()
         self.assertEqual(metadata.extra_data.get("hospital_repeat"), 3)
         self.assertEqual(metadata.extra_data.get("child_repeat"), 2)
-
-    def test_create_register_previous_candidates(self):
-        """Previous submissions are considered when creating repeat register"""
-        project = get_user_default_project(self.user)
-        md = """
-        | survey |
-        |        | type         | name            | label               |
-        |        | begin repeat | hospital_repeat |                     |
-        |        | text         | hospital        | Name of hospital    |
-        |        | begin repeat | child_repeat    |                     |
-        |        | text         | name            | Child's name        |
-        |        | decimal      | birthweight     | Child's birthweight |
-        |        | end_repeat   |                 |                     |
-        |        | end_repeat   |                 |                     |
-        | settings|             |                 |                     |
-        |         | form_title  | form_id         |                     |
-        |         | Births      | births          |                     |
-        """
-        self._publish_markdown(md, self.user, project)
-        xform = XForm.objects.all().order_by("-pk").first()
-        # Existing submission
-        xml = (
-            '<?xml version="1.0" encoding="UTF-8"?>'
-            '<data xmlns:jr="http://openrosa.org/javarosa" xmlns:orx='
-            '"http://openrosa.org/xforms" id="trees_update" version="2024050801">'
-            f"<formhub><uuid>{xform.uuid}</uuid></formhub>"
-            "<hospital_repeat>"
-            "<hospital>Aga Khan</hospital>"
-            "<child_repeat>"
-            "<name>Zakayo</name>"
-            "<birthweight>3.3</birthweight>"
-            "</child_repeat>"
-            "<child_repeat>"
-            "<name>Melania</name>"
-            "<birthweight>3.5</birthweight>"
-            "</child_repeat>"
-            "</hospital_repeat>"
-            "<hospital_repeat>"
-            "<hospital>Mama Lucy</hospital>"
-            "<child_repeat>"
-            "<name>Winnie</name>"
-            "<birthweight>3.1</birthweight>"
-            "</child_repeat>"
-            "</hospital_repeat>"
-            "<hospital_repeat>"
-            "<hospital>Nairobi West</hospital>"
-            "<child_repeat>"
-            "<name>Tom</name>"
-            "<birthweight>3.1</birthweight>"
-            "</child_repeat>"
-            "</hospital_repeat>"
-            "<meta>"
-            "<instanceID>uuid:cb5eb8fe-a046-4e75-9c7f-72183b871698</instanceID>"
-            "</meta>"
-            "</data>"
-        )
-        Instance.objects.create(xml=xml, user=self.user, xform=xform)
-        metadata = MetaData.objects.get(data_type="export_repeat_columns")
-        metadata.delete()  # Delete to simulate missing register
-
-        xml = (
-            '<?xml version="1.0" encoding="UTF-8"?>'
-            '<data xmlns:jr="http://openrosa.org/javarosa" xmlns:orx='
-            '"http://openrosa.org/xforms" id="trees_update" version="2024050801">'
-            f"<formhub><uuid>{xform.uuid}</uuid></formhub>"
-            "<hospital_repeat>"
-            "<hospital>Aga Khan</hospital>"
-            "<child_repeat>"
-            "<name>Zakayo</name>"
-            "<birthweight>3.3</birthweight>"
-            "</child_repeat>"
-            "<child_repeat>"
-            "<name>Melania</name>"
-            "<birthweight>3.5</birthweight>"
-            "</child_repeat>"
-            "</hospital_repeat>"
-            "<hospital_repeat>"
-            "<hospital>Mama Lucy</hospital>"
-            "<child_repeat>"
-            "<name>Winnie</name>"
-            "<birthweight>3.1</birthweight>"
-            "</child_repeat>"
-            "</hospital_repeat>"
-            "<meta>"
-            "<instanceID>uuid:45d27780-48fd-4035-8655-9332649385bd</instanceID>"
-            "</meta>"
-            "</data>"
-        )
-        Instance.objects.create(xml=xml, user=self.user, xform=xform)
-        metadata = MetaData.objects.get(data_type="export_repeat_columns")
-        # The previous submission should be considered since it has the most repeats
-        # for hospital_repeat
-        self.assertEqual(metadata.extra_data.get("hospital_repeat"), 3)
-        self.assertEqual(metadata.extra_data.get("child_repeat"), 2)
